Revert "feat[reservationPage] 이벤트 중복 방지"
This reverts commit c84631103a964350acc4239ebb32d836d2e00c79.

+++ frontend/js/reservation/reservastionList.js
@@ -40,10 +40,11 @@ document.addEventListener("DOMContentLoaded", function () {
 
 
   function displayPetModal(pets, date) {
-    const modal = createPetModal(pets.data);
+    const modal = createPetModal(pets.data);         
     modal.style.display = "block";
 
     modal.addEventListener("click", function (event) {
+      console.log(event.target)
       const selectedPetId = event.target.dataset.petId;
       if (selectedPetId) {
         createReservation({
@@ -119,6 +120,7 @@ document.addEventListener("DOMContentLoaded", function () {
       alert('예약을 취소하셨습니다.');
       getReservations();
     }).catch(error => {
+      
       console.error('예약 오류', error)
     });
   }
@@ -126,7 +128,7 @@ document.addEventListener("DOMContentLoaded", function () {
 
 
   function displayReservations(pets, reservations) {
-    const tableBody = document.querySelector("#reservationTable tbody")
+    const tableBody = document.querySelector("#reservationTable tbody");
 
     tableBody.innerHTML = "";
 
@@ -160,31 +162,24 @@ document.addEventListener("DOMContentLoaded", function () {
     }
 
     /**날짜 고르기 동작*/
-    let isEventRunning = false  /**이벤트 중복 방지, 이것 때문에 한참 해맸네 ㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋ*/
-
     tableBody.addEventListener("click", function (event) {
-      if (!isEventRunning) {
+      const clickedCell = event.target.closest("td");
+      const availability = (clickedCell.textContent)
+      const date = clickedCell.parentElement.cells[0].textContent;
 
-        isEventRunning = true;
-        console.log(isEventRunning)
-
-        const clickedCell = event.target.closest("td");
-        const availability = (clickedCell.textContent)
-        const date = clickedCell.parentElement.cells[0].textContent;
-
-        const reservation = reservations.find((r) => r.reservationDate === date);
-        if (clickedCell) {
-          if (!reservation && availability === '예약') {
-            displayPetModal(pets, date)
-          }
-          if (reservation && availability === 'cancel') {
-            const isConfirmed = confirm("예약을 취소하시겠습니까?");
-            if (isConfirmed) {
-              deleteReservation(reservation.reservationId)
-            }
-          }
+      const reservation = reservations.find((r) => r.reservationDate === date);
+      console.log(clickedCell)
+      if (clickedCell) { 
+        if (!reservation && availability === '예약') {
+          displayPetModal(pets, date)
+        }
+        if (reservation && availability === 'cancel') {
+          const isConfirmed = confirm("예약을 취소하시겠습니까?");
+          if (isConfirmed) deleteReservation(reservation.reservationId)
         }
       }
+
+
     });
   }
   getReservations();
